ref: consolidate query functions in their own file
