+++ contracts/hub/stable_vlp/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod contract;
-pub mod math;
 pub mod migrate;
 pub mod query;
 pub mod reply;
