+++ contracts/hub/stable_vlp/src/query.rs
@@ -1,186 +1,1 @@
-use cosmwasm_std::{ensure, to_json_binary, Binary, Decimal, Decimal256, Deps, Env, Uint128};
-use euclid::chain::ChainUid;
-use euclid::error::ContractError;
-use euclid::msgs::stable_vlp::{
-    AllStablePoolsResponse, FeeResponse, GetLiquidityResponse, GetStateResponse, GetSwapResponse,
-    StablePoolInfo, StablePoolResponse, TotalFeesPerDenomResponse, TotalFeesResponse,
-};
-use euclid::pool::{PoolConfig, DEFAULT_AMP_FACTOR};
-use euclid::swap::NextSwapVlp;
-use euclid::token::Token;
-use euclid::utils::math::Decimal256Ext;
 
-use crate::math::compute_swap;
-use crate::state::{AMP_FACTOR, BALANCES, CHAIN_LP_TOKENS, STATE};
-use euclid::pool::State;
-// Function to simulate swap in a query
-pub fn query_simulate_swap(
-    deps: Deps,
-    asset_in: Token,
-    amount_in: Uint128,
-    next_swaps: Vec<NextSwapVlp>,
-) -> Result<Binary, ContractError> {
-    // Verify that the asset amount is non-zero
-    ensure!(!amount_in.is_zero(), ContractError::ZeroAssetAmount {});
-
-    let state = STATE.load(deps.storage)?;
-
-    let pair = state.pair.clone();
-
-    // asset should match either token
-    ensure!(asset_in.exists(pair), ContractError::AssetDoesNotExist {});
-
-    // Get Fee from the state
-    let fee = state.clone().fee;
-
-    let lp_fee = amount_in.checked_mul_floor(Decimal::bps(fee.lp_fee_bps))?;
-    let euclid_fee = amount_in.checked_mul_floor(Decimal::bps(fee.euclid_fee_bps))?;
-
-    // Calcuate the sum of fees
-    let total_fee = lp_fee.checked_add(euclid_fee)?;
-
-    // Calculate the amount of asset to be swapped
-    let swap_amount = amount_in.checked_sub(total_fee)?;
-
-    let asset_out = state.pair.get_other_token(asset_in.clone());
-
-    let token_in_reserve = BALANCES.load(deps.storage, asset_in)?;
-    let token_out_reserve = BALANCES.load(deps.storage, asset_out.clone())?;
-    let amp_factor = AMP_FACTOR.load(deps.storage).unwrap_or(DEFAULT_AMP_FACTOR);
-    let receive_amount = compute_swap(
-        &Decimal256::from_integer(swap_amount),
-        &Decimal256::from_integer(token_in_reserve),
-        &Decimal256::from_integer(token_out_reserve),
-        amp_factor,
-    )?;
-    let response = match next_swaps.split_first() {
-        Some((next_swap, forward_swaps)) => {
-            let next_swap_response: GetSwapResponse = deps.querier.query_wasm_smart(
-                next_swap.vlp_address.clone(),
-                &euclid::msgs::stable_vlp::QueryMsg::SimulateSwap {
-                    asset: asset_out,
-                    asset_amount: receive_amount.return_amount,
-                    swaps: forward_swaps.to_vec(),
-                },
-            )?;
-            Ok(to_json_binary(&next_swap_response)?)
-        }
-        None => Ok(to_json_binary(&GetSwapResponse {
-            amount_out: receive_amount.return_amount,
-            asset_out,
-            spread_amount: receive_amount.spread_amount,
-        })?),
-    };
-    response
-}
-
-// Function to query the total liquidity
-pub fn query_liquidity(deps: Deps, _env: Env) -> Result<Binary, ContractError> {
-    let state = STATE.load(deps.storage)?;
-    let pair = state.pair.clone();
-    Ok(to_json_binary(&GetLiquidityResponse {
-        pair,
-        token_1_reserve: BALANCES
-            .may_load(deps.storage, state.pair.token_1)?
-            .unwrap_or_default(),
-        token_2_reserve: BALANCES
-            .may_load(deps.storage, state.pair.token_2)?
-            .unwrap_or_default(),
-        total_lp_tokens: state.total_lp_tokens,
-    })?)
-}
-
-// Function to query fee of the contract
-pub fn query_fee(deps: Deps) -> Result<Binary, ContractError> {
-    let state = STATE.load(deps.storage)?;
-    Ok(to_json_binary(&FeeResponse { fee: state.fee })?)
-}
-
-// Function to query total fees collected of the contract
-pub fn query_total_fees_collected(deps: Deps) -> Result<Binary, ContractError> {
-    let state = STATE.load(deps.storage)?;
-    Ok(to_json_binary(&TotalFeesResponse {
-        total_fees: state.total_fees_collected,
-    })?)
-}
-
-pub fn query_total_fees_per_denom(deps: Deps, denom: String) -> Result<Binary, ContractError> {
-    let total_fees_collected = STATE.load(deps.storage)?.total_fees_collected;
-
-    let lp_fees = total_fees_collected.lp_fees.get_fee(denom.as_str());
-    let euclid_fees = total_fees_collected.euclid_fees.get_fee(denom.as_str());
-
-    Ok(to_json_binary(&TotalFeesPerDenomResponse {
-        lp_fees,
-        euclid_fees,
-    })?)
-}
-
-pub fn query_state(deps: Deps) -> Result<Binary, ContractError> {
-    let state = STATE.load(deps.storage)?;
-    Ok(to_json_binary(&GetStateResponse {
-        pair: state.pair,
-        router: state.router,
-        virtual_balance: state.virtual_balance,
-        fee: state.fee,
-        total_fees_collected: state.total_fees_collected,
-        last_updated: state.last_updated,
-        total_lp_tokens: state.total_lp_tokens,
-        admin: state.admin,
-        pool_config: PoolConfig::Stable {
-            amp_factor: AMP_FACTOR.may_load(deps.storage)?,
-        },
-    })?)
-}
-
-// Function to query a Euclid Pool Information for this pair
-pub fn query_pool(deps: Deps, chain_uid: ChainUid) -> Result<Binary, ContractError> {
-    let state = STATE.load(deps.storage)?;
-
-    let chain_lp_tokens = CHAIN_LP_TOKENS.load(deps.storage, chain_uid)?;
-
-    let reserve_1 = BALANCES.load(deps.storage, state.pair.token_1.clone())?;
-
-    let reserve_2 = BALANCES.load(deps.storage, state.pair.token_2.clone())?;
-
-    let pool = get_pool(&state, chain_lp_tokens, reserve_1, reserve_2)?;
-
-    Ok(to_json_binary(&pool)?)
-}
-// Function to query all Euclid Pool Information
-pub fn query_all_pools(deps: Deps) -> Result<Binary, ContractError> {
-    let state = STATE.load(deps.storage)?;
-
-    let reserve_1 = BALANCES.load(deps.storage, state.pair.token_1.clone())?;
-
-    let reserve_2 = BALANCES.load(deps.storage, state.pair.token_2.clone())?;
-
-    let pools: Result<_, ContractError> = CHAIN_LP_TOKENS
-        .range(deps.storage, None, None, cosmwasm_std::Order::Ascending)
-        .map(|item| {
-            let (chain_uid, chain_lp_tokens) = item?;
-            let pool = get_pool(&state, chain_lp_tokens, reserve_1, reserve_2)?;
-
-            Ok::<StablePoolInfo, ContractError>(StablePoolInfo { chain_uid, pool })
-        })
-        .collect();
-
-    Ok(to_json_binary(&AllStablePoolsResponse { pools: pools? })?)
-}
-
-fn get_pool(
-    state: &State,
-    chain_lp_tokens: Uint128,
-    reserve_1: Uint128,
-    reserve_2: Uint128,
-) -> Result<StablePoolResponse, ContractError> {
-    Ok(StablePoolResponse {
-        reserve_1: reserve_1
-            .checked_multiply_ratio(chain_lp_tokens, state.total_lp_tokens)
-            .unwrap_or(Uint128::zero()),
-        reserve_2: reserve_2
-            .checked_multiply_ratio(chain_lp_tokens, state.total_lp_tokens)
-            .unwrap_or(Uint128::zero()),
-        lp_shares: chain_lp_tokens,
-    })
-}
